// src/lib/api/models/BookableResourceAvailability.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Firefly API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { AvailabilitySlot } from './AvailabilitySlot';
import {
    AvailabilitySlotFromJSON,
    AvailabilitySlotFromJSONTyped,
    AvailabilitySlotToJSON,
    AvailabilitySlotToJSONTyped,
} from './AvailabilitySlot';

/**
 * 
 * @export
 * @interface BookableResourceAvailability
 */
export interface BookableResourceAvailability {
    /**
     * 
     * @type {string}
     * @memberof BookableResourceAvailability
     */
    providerId: string;
    /**
     * 
     * @type {string}
     * @memberof BookableResourceAvailability
     */
    providerName: string;
    /**
     * 
     * @type {Array<AvailabilitySlot>}
     * @memberof BookableResourceAvailability
     */
    slots: Array<AvailabilitySlot>;
}

/**
 * Check if a given object implements the BookableResourceAvailability interface.
 */
export function instanceOfBookableResourceAvailability(value: object): value is BookableResourceAvailability {
    if (!('providerId' in value) || value['providerId'] === undefined) return false;
    if (!('providerName' in value) || value['providerName'] === undefined) return false;
    if (!('slots' in value) || value['slots'] === undefined) return false;
    return true;
}

export function BookableResourceAvailabilityFromJSON(json: any): BookableResourceAvailability {
    return BookableResourceAvailabilityFromJSONTyped(json, false);
}

export function BookableResourceAvailabilityFromJSONTyped(json: any, ignoreDiscriminator: boolean): BookableResourceAvailability {
    if (json == null) {
        return json;
    }
    return {
        
        'providerId': json['provider_id'],
        'providerName': json['provider_name'],
        'slots': ((json['slots'] as Array<any>).map(AvailabilitySlotFromJSON)),
    };
}

export function BookableResourceAvailabilityToJSON(json: any): BookableResourceAvailability {
    return BookableResourceAvailabilityToJSONTyped(json, false);
}

export function BookableResourceAvailabilityToJSONTyped(value?: BookableResourceAvailability | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'provider_id': value['providerId'],
        'provider_name': value['providerName'],
        'slots': ((value['slots'] as Array<any>).map(AvailabilitySlotToJSON)),
    };
}