// src/lib/api/models/AvailabilitySlot.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Firefly API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AvailabilitySlot
 */
export interface AvailabilitySlot {
    /**
     * 
     * @type {string}
     * @memberof AvailabilitySlot
     */
    start: string;
    /**
     * 
     * @type {string}
     * @memberof AvailabilitySlot
     */
    end: string;
}

/**
 * Check if a given object implements the AvailabilitySlot interface.
 */
export function instanceOfAvailabilitySlot(value: object): value is AvailabilitySlot {
    if (!('start' in value) || value['start'] === undefined) return false;
    if (!('end' in value) || value['end'] === undefined) return false;
    return true;
}

export function AvailabilitySlotFromJSON(json: any): AvailabilitySlot {
    return AvailabilitySlotFromJSONTyped(json, false);
}

export function AvailabilitySlotFromJSONTyped(json: any, ignoreDiscriminator: boolean): AvailabilitySlot {
    if (json == null) {
        return json;
    }
    return {
        
        'start': json['start'],
        'end': json['end'],
    };
}

export function AvailabilitySlotToJSON(json: any): AvailabilitySlot {
    return AvailabilitySlotToJSONTyped(json, false);
}

export function AvailabilitySlotToJSONTyped(value?: AvailabilitySlot | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'start': value['start'],
        'end': value['end'],
    };
}